[fix] Serializers type was coming wrong

+++ app/serializers.py
@@ -4,7 +4,7 @@ class EquitySerializer(serializers.Serializer):
     code = serializers.CharField(max_length= 6)
     name = serializers.CharField(max_length= 26)
     # Storing as Integer for future usage
-    open_value = serializers.IntegerField()
-    high_value = serializers.IntegerField()
-    low_value = serializers.IntegerField()
-    close_value = serializers.IntegerField()
+    open_value = serializers.FloatField()
+    high_value = serializers.FloatField()
+    low_value = serializers.FloatField()
+    close_value = serializers.FloatField()
